map_outlet_ids {hyRefactor}	R Documentation

Map outlets from COMID to ID for aggregate catchments

Description

given reconciled flowlines and a set of source outlets, returns a set of outlets with reconciled IDs suitable for use with aggregate_catchments.

Usage

map_outlet_ids(source_outlets, reconciled)

Arguments

source_outlets
data.frame with COMID and type columns

reconciled
data.frame as returned by refactor workflow

[Package hyRefactor version 0.4.6 Index]