reconcile_catchment_divides {hyRefactor}	R Documentation

Reconcile Catchment Divides

Description

Reconciles catchment divides according to the output of reconcile_collapsed_flowlines and refactor_nhdplus

Usage

reconcile_catchment_divides(
  catchment,
  fline_ref,
  fline_rec,
  fdr = NULL,
  fac = NULL,
  para = 2,
  cache = NULL,
  min_area_m = 800,
  snap_distance_m = 100,
  simplify_tolerance_m = 40,
  vector_crs = NULL
)

Arguments

catchment
sf data.frame NHDPlus Catchment or CatchmentSP layers for included COMIDs

fline_ref
sf data.frame flowlines as returned by refactor_nhdplus and reconcile_collapsed_flowlines

fline_rec
sf data.frame flowpaths as returned by reconcile_collapsed_flowlines

fdr
raster D8 flow direction

fac
raster flow accumulation

para
integer numer of cores to use for parallel execution

cache
path .rda to cache incremental outputs

min_area_m
minimum area in m^2 to filter out slivers (caution, use with care!!)

snap_distance_m
distance in meters to snap raster generated geometry to polygon geometry

simplify_tolerance_m
dTolerance in meters for simplification of grid-cell based polygons

vector_crs
any object compatible with sf::st_crs. Used for vector-based calculations in case that raster projection is not suitable (e.g. lon/lat) – must result in units of meters.

Details

Note that all inputs must be passed in the same projection.

Value

Catchment divides that have been split and collapsed according to input flowpaths

See Also

The refactor_nhdplus function implements a complete workflow using 'reconcile_collapsed_flowlines()' and can be used in prep for this function.

[Package hyRefactor version 0.4.6 Index]