reconcile_collapsed_flowlines {hyRefactor}	R Documentation

Reconcile Collapsed Flowlines

Description

Reconciles output of collapse_flowlines giving a unique ID to each new flowpath and providing a mapping to NHDPlus COMIDs.

Usage

reconcile_collapsed_flowlines(flines, geom = NULL, id = "COMID")

Arguments

flines
data.frame with COMID, toCOMID, LENGTHKM, LevelPathI, Hydroseq, and TotDASqKM columns

geom
sf data.frame for flines

id
character id collumn name.

Value

reconciled flowpaths with new ID, toID, LevelPathID, and Hydroseq identifiers. Note that all the identifiers are new integer IDs. LevelPathID and Hydroseq are consistent with the LevelPathID and Hydroseq from the input NHDPlus flowlines.

See Also

The refactor_nhdplus function implements a complete workflow using 'reconcile_collapsed_flowlines()'.

[Package hyRefactor version 0.4.6 Index]