Reconcile Catchment Divides

Reconciles catchment divides according to the output of reconcile_collapsed_flowlines and refactor_nhdplus

reconcile_catchment_divides(
  catchment,
  fline_ref,
  fline_rec,
  fdr,
  fac,
  para = 2,
  cache = NULL
)

Arguments

catchment
sf data.frame NHDPlus Catchment or CatchmentSP layers for included COMIDs

fline_ref
sf data.frame flowlines as returned by refactor_nhdplus and reconcile_collapsed_flowlines

fline_rec
sf data.frame flowpaths as returned by reconcile_collapsed_flowlines

fdr
raster D8 flow direction

fac
raster flow accumulation

para
integer numer of cores to use for parallel execution

cache
path .rda to cache incremental outputs

Value

Catchment divides that have been split and collapsed according to input flowpaths

Details

Note that all inputs must be passed in the same projection.

See also